pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astro-mcp"
version = "1.0.0"
requires-python = ">=3.11"
description = "Astrological MCP Server — high-precision astrology tools for LLM agents"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "mcp>=1.0.0",
    "pyswisseph>=2.10.3.2",
    "geopy>=2.4.0",
    "timezonefinder>=6.5.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
astro-mcp = "astro_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
